+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-04 13:29:32
-// Commit: test(ui): implement widget integration tests for ui
+// Updated at: 2026-04-04 16:30:22
+// Commit: fix(ui): fix memory leak by cancelling ui subscription
 
 class uiComponentState {
-  final String commitRef = '107';
-  final String lastMessage = 'test(ui): implement widget integration tests for ui';
-  final DateTime timestamp = DateTime.parse('2026-04-04T13:29:32+05:00');
+  final String commitRef = '118';
+  final String lastMessage = 'fix(ui): fix memory leak by cancelling ui subscription';
+  final DateTime timestamp = DateTime.parse('2026-04-04T16:30:22+05:00');
 }